Use a logo with correct background colours
The logo used before had a transparent region within the SR logo, which
looked terrible on the stream. I have also swapped the logo for a SVG
version of the image and fixed the size of the img to be the same as the
dimensions of the previous PNG.

The SVG is from https://github.com/srobo/style/pull/33 rather than the main
branch of the style repo.